-- profile -- app/Http/Controllers/HumanResources/Profile/ProfileController.php -- resources/views/humanresources/profile/edit.blade.php

+++ app/Http/Controllers/HumanResources/Profile/ProfileController.php
@@ -10,6 +10,8 @@ use Illuminate\Http\RedirectResponse;
 // load models
 use App\Models\Staff;
 use App\Models\HumanResources\HREmergency;
+use App\Models\HumanResources\HRStaffSpouse;
+use App\Models\HumanResources\HRStaffChildren;
 
 // load validation
 use App\Http\Requests\HumanResources\Profile\ProfileRequestUpdate;
@@ -65,28 +67,81 @@ class ProfileController extends Controller
     return view('humanresources.profile.edit', compact('profile'));
   }
 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
   /**
    * Update the specified resource in storage.
    */
   public function update(ProfileRequestUpdate $request, Staff $profile): RedirectResponse
   {
     // return $request->emer;
-
     // return \Carbon\Carbon::parse($request->dob)->format('Y-m-d');
 
     $profile->update($request->only(['ic', 'mobile', 'email', 'address', 'dob', 'gender_id', 'nationality_id', 'race_id', 'religion_id', 'marital_status_id']));
 
-    foreach ($request->emer as $value) {
+    foreach ($request->emer as $emer_value) {
       $HREmergency = HREmergency::updateOrCreate(
         [
-          'id' => $value['id']
+          'id' => $emer_value['id']
         ],
         [
-          'staff_id' => $value['staff_id'],
-          'contact_person' => $value['contact_person'],
-          'phone' => $value['phone'],
-          'address' => $value['address'],
-          'relationship_id' => $value['relationship_id'],
+          'staff_id' => $emer_value['staff_id'],
+          'contact_person' => $emer_value['contact_person'],
+          'phone' => $emer_value['phone'],
+          'address' => $emer_value['address'],
+          'relationship_id' => $emer_value['relationship_id'],
+        ]
+      );
+    }
+
+    foreach ($request->spou as $spou_value) {
+      $HRStaffSpouse = HRStaffSpouse::updateOrCreate(
+        [
+          'id' => $spou_value['id']
+        ],
+        [
+          'staff_id' => $spou_value['staff_id'],
+          'spouse' => $spou_value['spouse'],
+          'id_card_passport' => $spou_value['id_card_passport'],
+          'phone' => $spou_value['phone'],
+          'dob' => $spou_value['dob'],
+          'profession' => $spou_value['profession'],
+        ]
+      );
+    }
+
+    foreach ($request->chil as $chil_value) {
+      $HRStaffChildren = HRStaffChildren::updateOrCreate(
+        [
+          'id' => $chil_value['id']
+        ],
+        [
+          'staff_id' => $chil_value['staff_id'],
+          'children' => $chil_value['children'],
+          'dob' => $chil_value['dob'],
+          'gender_id' => $chil_value['gender_id'],
+          'health_status_id' => $chil_value['health_status_id'],
+          'education_level_id' => $chil_value['education_level_id'],
         ]
       );
     }
@@ -95,20 +150,86 @@ class ProfileController extends Controller
     return Redirect::route('profile.show', $profile);
   }
 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
   /**
    * Remove the specified resource from storage.
    */
-  public function destroy(Staff $profile)
+  public function destroy(Request $request, Staff $profile)
   {
-    $HREmergency = HREmergency::destroy(
-      [
-        'id' => $profile['id']
-      ]
-    );
+    if ($request->table == 'emergency') {
+      $HREmergency = HREmergency::destroy(
+        [
+          'id' => $profile['id']
+        ]
+      );
 
-    return response()->json([
-      'status' => 'success',
-      'message' => 'Your emergency contact has been deleted.',
-    ]);
+      return response()->json([
+        'status' => 'success',
+        'message' => 'Your emergency contact has been deleted.',
+      ]);
+    }
+
+
+    if ($request->table == 'spouse') {
+      $HRStaffSpouse = HRStaffSpouse::destroy(
+        [
+          'id' => $profile['id']
+        ]
+      );
+
+      return response()->json([
+        'status' => 'success',
+        'message' => 'Your spouse contact has been deleted.',
+      ]);
+    }
+
+
+    if ($request->table == 'children') {
+      $HRStaffChildren = HRStaffChildren::destroy(
+        [
+          'id' => $profile['id']
+        ]
+      );
+
+      return response()->json([
+        'status' => 'success',
+        'message' => 'Your children contact has been deleted.',
+      ]);
+    }
   }
 }
